requirements.txt
branca==0.3.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
decorator==4.2.1
-e git+git@github.com:python-visualization/folium.git@c228c261be42d801809e0ba037dbe14b8229fb4b#egg=folium
GeoAlchemy2==0.4.2
geojson==2.3.0
geojsonio==0.0.3
github3.py==1.1.0
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
MarkupSafe==1.0
munch==2.3.2
numpy==1.14.1
pandas==0.22.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pyproj==1.9.5.1
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.5
tornado==5.0.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
